Carrito elias

+++ VentasCarrito/VentasCarrito/Modelos/ProductoApi.cs
@@ -6,6 +6,6 @@
         public string nombre_producto { get; set; }
         public int existencia { get; set; }
         public string marca { get; set; }
-        public decimal precio { get; set; }
+        public double precio { get; set; }
     }
 }
